requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.20.8
botocore==1.23.9
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-allauth==0.45.0
django-braces==1.15.0
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dynaconf==3.1.7
hvac==0.11.2
idna==3.3
install==1.3.4
Jinja2==3.0.3
jmespath==0.10.0
jwcrypto==1.0
knox==0.1.14
kombu==5.2.2
loguru==0.5.3
MarkupSafe==2.0.1
oauthlib==3.1.1
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
